requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase>=2.10.0
httpx>=0.26.0
websockets>=14.0,<16
python-dotenv==1.0.1
